Aloha

+++ function_pointers/1-array_iterator.c
@@ -1,6 +1,7 @@
 #include "function_pointers.h"
 #include<stdio.h>
-#include<stddef.h>
+#include <stddef.h>
+#include <stddef.h>
 
 /**
  * array_iterator - executes a function given as a parameter
@@ -11,7 +12,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	if (array != NULL && action != NULL)
 	{
